refactor: remove unused vms references and update related code for clarity
